initialized several sections
I have written something to get started in every section of the DesignDocument.md that did not previously have something in it except Program Organization, Major Classes, and General Architectural Quality. I think the bulk of the important design information will be in these sections, and the rest are more edge cases to think about in case they come up later.

In the morning, feel free to proofread, add to, and revise as desired. I'm most unsure of what else to put for the Fault Tolerance, Architectural Feasibility, and Overengineering sections. Although, maybe i'm overthinking this and these may be more minor sections where what I have now is more than enough