Added buttons to pause menu
added buttons to pause menu. Apparently the pause menu canvas needs a different display mode than what the current HUD canvas has, so I had to use a second canvas and make a prefab out of it. I also added to the restart button script so the R key will restart the level as well as the button on screen.